[enh] use FR/EN message, when local node is detected
[fix] Fix windows build
[enh] Starting an OSX build

+++ install.sh
@@ -3,12 +3,14 @@
 ROOT_DIR=$PWD
 TEMP_DIR=$PWD/tmp
 NV_VERSION=0.35.2
+CHROMIUM_MAJOR_VERSION=71
+CESIUM_DEFAULT_VERSION=1.2.9
 
 # Check first arguments = version
 if [[ $1 =~ ^[0-9]+.[0-9]+.[0-9]+((a|b)[0-9]+)?$ ]];
 then
   VERSION="$1"
-  echo "Using given version: $VERSION"
+  echo "Using Cesium version: $VERSION"
 else
   if [[ -f $ROOT_DIR/src/nw/cesium/manifest.json ]];
   then
@@ -17,14 +19,15 @@ else
 
   if [[ $VERSION =~ ^[0-9]+.[0-9]+.[0-9]+((a|b)[0-9]+)?$ ]];
   then
-    echo "Using detected version: $VERSION"
+    echo "Using Cesium detected version: $VERSION"
   else
-    VERSION="1.0.0";
-    echo "No version detected. Using default: $VERSION"
+    VERSION="${CESIUM_DEFAULT_VERSION}";
+    echo "No Cesium version detected. Using default: $VERSION"
   fi
 fi
 
 mkdir -p $TEMP_DIR && cd $TEMP_DIR
+
 
 # Install NW.js
 if [[ ! -f $ROOT_DIR/src/nw/nw ]];
@@ -35,6 +38,18 @@ then
   rm nwjs-sdk-v${NV_VERSION}-linux-x64.tar.gz
   rmdir nwjs-sdk-v${NV_VERSION}-linux-x64
   rmdir nw
+
+# Check NW version
+else
+  cd ${ROOT_DIR}/src/nw
+  NW_ACTUAL_VERSION=`./nw --version | grep nwjs | awk '{print $2}'`
+  echo "Using Chromium version: ${NW_ACTUAL_VERSION}"
+  CHROMIUM_ACTUAL_MAJOR_VERSION=`echo ${NW_ACTUAL_VERSION} | awk '{split($0, array, ".")} END{print array[1]}'`
+  cd ${ROOT_DIR}
+  if [[ ${CHROMIUM_ACTUAL_MAJOR_VERSION} -ne ${CHROMIUM_MAJOR_VERSION} ]]; then
+    echo "Bad Chromium major version: ${CHROMIUM_ACTUAL_MAJOR_VERSION}. Expected version ${CHROMIUM_MAJOR_VERSION}"
+    exit -1
+  fi
 fi
 
 # Remove old Cesium version
@@ -57,12 +72,20 @@ fi
 if [[ ! -f $ROOT_DIR/src/nw/cesium/index.html ]]; then
     echo "Downloading Cesium ${VERSION}..."
 
-    mkdir cesium_unzip && cd cesium_unzip
+    mkdir -p ${TEMP_DIR}/cesium_unzip
     wget "https://github.com/duniter/cesium/releases/download/v${VERSION}/cesium-v${VERSION}-web.zip"
-    unzip "cesium-v${VERSION}-web.zip"
+    if [[ ! $? -eq 0 ]]; then
+        echo "Could not download Cesium web release !"
+        exit -1;
+    fi
+    unzip "cesium-v${VERSION}-web.zip" -d ${TEMP_DIR}/cesium_unzip
     rm "cesium-v${VERSION}-web.zip"
+
+    # Add node.js file into HTML files
+    cd ${TEMP_DIR}/cesium_unzip
     sed -i 's/<script src="config.js"><\/script>/<script src="config.js"><\/script><script src="node.js"><\/script>/' "index.html"
     sed -i 's/<script src="config.js"><\/script>/<script src="config.js"><\/script><script src="node.js"><\/script>/' "debug.html"
+
     mv * "$ROOT_DIR/src/nw/cesium/"
     cd ..
     rmdir cesium_unzip
